+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [0.2.86] - 2018-08-17
+### Added
+- The `js show if` and `js hide if` field modifiers.
+- The `continue button field` feature.
+### Changed
+- Upgraded Font Awesome to 5.2.0.
+- Two `DAList`s added together will yield a `DAList`.
+### Fixed
+- `zip_file()` returned an empty archive when given a
+  `DAFileCollection`.
+
 ## [0.2.85] - 2018-08-15
 ### Fixed
 - Problem with screen reader flagging help text with Markdown